added null checks to repository

+++ VehicleAPI/Services/VehicleRepository.cs
@@ -6,6 +6,7 @@ using VehicleAPI.VehicleDb;
 using VehicleAPI.Data.SeedData;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.EntityFrameworkCore;
+using System.Linq.Expressions;
 
 namespace VehicleAPI.Services
 {
@@ -19,34 +20,60 @@ namespace VehicleAPI.Services
         }
         public async Task DeleteCar(int id)
         {
-            var car = await _context.Cars.FindAsync(id);
-            _context.Cars.Remove(car);
-            await _context.SaveChangesAsync();
+                var car = await _context.Cars.FindAsync(id);
+
+                if (car == null)
+                {
+                    return null;
+                }
+                _context.Cars.Remove(car);
+                await _context.SaveChangesAsync();
         }
 
         public async Task<Car> GetCarById(int id)
         {
             var car = await _context.Cars.FindAsync(id);
+
+            if(car is null)
+            {
+                return null;
+            }
+
             return car;
         }
 
         public async Task<Car> TurnHeadlights(int id, bool state)
         {
             var car = await _context.Cars.FindAsync(id);
-            car.headlightsOn = state;
+            
+            if (car is null)
+            {
+                return null;
+            }
+
+            car.HeadlightsOn = state;
             await _context.SaveChangesAsync();
             return car;
         }
 
         public async Task<IEnumerable<Boat>> SelectBoatsByColor(string color)
         {
+            if (string.IsNullOrWhiteSpace(color))
+            {
+                return new List<Boat>();
+            }
+
             return await _context.Boats
-                .Where(bo => bo.Color.ToLower() == color.ToLower())
-                .ToListAsync();
+                 .Where(bo => bo.Color.ToLower() == color.ToLower())
+                 .ToListAsync();
         }
 
         public async Task<IEnumerable<Bus>> SelectBusesByColor(string color)
         {
+            if (string.IsNullOrWhiteSpace(color))
+            {
+                return new List<Bus>();
+            }
             return await _context.Buses
                 .Where(bu => bu.Color.ToLower() == color.ToLower())
                 .ToListAsync();
@@ -54,6 +81,10 @@ namespace VehicleAPI.Services
 
         public async Task<IEnumerable<Car>> SelectCarsByColor(string color)
         {
+            if (string.IsNullOrWhiteSpace(color))
+            {
+                return new List<Car>();
+            }
             return await _context.Cars
                 .Where(c => c.Color.ToLower() == color.ToLower())
                 .ToListAsync();
